+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Clean unremoved arguments
+- Overlapping tutorial and header
 
 ## 2.2.0 - 2020-07-04
 ### Added